Updated Core HTML
I updated the html to use a bootstrap layout. The recipe page has a
basic layout and accounts for an image when we implement that. I also
updated the index page so that it now has a similar layout, although it
currently does not have a recipe per page limit.